broke filters out into a submodule
